Move view ownership to ChunkStore, eliminating reverse lookups
